+++ app/src/main/java/org/aquamarine5/brainspark/chaoxingsignfaker/api/ChaoxingRecommendHelper.kt
@@ -37,7 +37,7 @@ object ChaoxingRecommendHelper {
         }
 
 
-    suspend fun getRecommendedCourses(context: Context): List<Pair<Int, Int>> {
+    suspend fun getRecommendedCourses(context: Context): List<Pair<Int, Long>> {
         val allHabits = getAllLearnedHabits(context)
         val now = LocalDateTime.now()
         val currentDayOfWeek = now.dayOfWeek.value
@@ -79,7 +79,7 @@ object ChaoxingRecommendHelper {
 
     suspend fun analyseRecommendHabit(
         classId: Int,
-        courseId: Int,
+        courseId: Long,
         recommendRecords: Map<Int, RecommendRecordList>,
         recommendHabits: List<RecommendHabit>,
         currentRecommendRecord: RecommendRecord,
@@ -135,7 +135,7 @@ object ChaoxingRecommendHelper {
     suspend fun recordRecommendEvent(
         context: Context,
         classId: Int,
-        courseId: Int,
+        courseId: Long,
         client: ChaoxingHttpClient
     ) =
         withContext(Dispatchers.IO) {
