+++ app/src/main/java/org/aquamarine5/brainspark/chaoxingsignfaker/api/SignDestination.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.Immutable
 interface SignDestination {
     val activeId: Long
     val classId: Int
-    val courseId: Int
+    val courseId: Long
     val endTime: Long?
     val isCloneSession: Boolean
 }
